Split private swap actions into separate modules
