+++ lib/ex_okex/swap/private.ex
@@ -5,13 +5,11 @@ defmodule ExOkex.Swap.Private do
   [API docs](https://www.okex.com/docs/en/#swap-README)
   """
 
-  import ExOkex.Api.Private
+  alias ExOkex.Swap.Private
 
   @type params :: map
   @type config :: ExOkex.Config.t()
   @type response :: ExOkex.Api.response()
-
-  @prefix "/api/swap/v3"
 
   @doc """
   Place a new order.
@@ -31,9 +29,7 @@ defmodule ExOkex.Swap.Private do
   {:ok, %{"order_info" => [%{"error_code" => 0, "error_message" => "", "order_id" => "2653481276189696"}], "result" => true}}
   """
   @spec create_order(params, config | nil) :: response
-  def create_order(params, config \\ nil) do
-    post("#{@prefix}/orders", params, config)
-  end
+  defdelegate create_order(params, config \\ nil), to: Private.CreateOrder
 
   @doc """
   Place multiple orders for specific trading pairs (up to 4 trading pairs, maximum 4 orders each)
@@ -51,15 +47,13 @@ defmodule ExOkex.Swap.Private do
       "leverage":"10" },
     ])
 
-    # TODO: Add response sample
-
   """
   @spec create_bulk_orders([params], config | nil) :: response
-  def create_bulk_orders(params, config \\ nil) do
-    post("#{@prefix}/orders", params, config)
-  end
+  defdelegate create_bulk_orders(params, config \\ nil), to: Private.CreateBulkOrders
 
-  defdelegate create_batch_orders(params, config \\ nil), to: __MODULE__, as: :create_bulk_orders
+  defdelegate create_batch_orders(params, config \\ nil),
+    to: Private.CreateBulkOrders,
+    as: :create_bulk_orders
 
   @doc """
   Cancelling an unfilled order.
@@ -70,12 +64,9 @@ defmodule ExOkex.Swap.Private do
 
       iex> ExOkex.Swap.cancel_orders("BTC-USD-180309", [1600593327162368,1600593327162369])
 
-      # TODO: Add response
   """
-  def cancel_orders(instrument_id, order_ids \\ [], params \\ %{}, config \\ nil) do
-    new_params = Map.merge(params, %{ids: order_ids})
-    post("#{@prefix}/cancel_batch_orders/#{instrument_id}", new_params, config)
-  end
+  defdelegate cancel_orders(instrument_id, order_ids \\ [], params \\ %{}, config \\ nil),
+    to: Private.CancelOrders
 
   @doc """
   Get the swap account info of all token.
@@ -86,11 +77,8 @@ defmodule ExOkex.Swap.Private do
 
       iex(3)> ExOkex.Swap.list_accounts()
 
-      # TODO: Add Response
   """
-  def list_accounts(config \\ nil) do
-    get("#{@prefix}/accounts", %{}, config)
-  end
+  defdelegate list_accounts(config \\ nil), to: Private.ListAccounts
 
   @doc """
   Get the information of holding positions of a contract.
@@ -102,7 +90,5 @@ defmodule ExOkex.Swap.Private do
       iex(3)> ExOkex.Swap.get_position("BTC-USD-190329")
 
   """
-  def get_position(instrument_id, config \\ nil) do
-    get("#{@prefix}/#{instrument_id}/position", %{}, config)
-  end
+  defdelegate position(instrument_id, config \\ nil), to: Private.Position
 end
